Refactor controllers into api package
