+++ apps/api/anyset/api/__init__.py
@@ -1,0 +1,5 @@
+"""Anyset API package."""
+
+from .routes import router
+
+__all__ = ["router"]
