+++ apps/api/anyset/api/models.py
@@ -1,0 +1,1 @@
+"""Models for the AnySet API."""
